add compile-css and build:css command
